Continuing to work on this.
I'm sure it's a song you've heard before, but this assignment really *wasn't* completed last-minute...I have been working on this since last Wednesday, and unfortunately am concerned I haven't made much progress. I have consulted with several Python professionals, but the ability to code in Python =/= the ability to teach in Python, so their advanced code is of little utility to my understanding of the core concepts at stake here. In the meantime, I'll continue at it - and hopefully get this solved.

I am recruiting one more person for help, so I hope they'll be able to offer some pointers (as I flip back and forth between the notes, StackOverflow, and this code!).